:bug: Move StateMachineService local Agent to a new interface
- Resolve an issue with unit test delegates being cached, which interfered with scenario tests
- Add IStateMachineServiceLocalAdapter and implementation which support clearer mocking and better focuses StateMachineService
- Remove regions from StateMachineService
- Make non-virtual internal methods private on StateMachineService
- Closes #8
