+++ src/REstate.Remote/Services/IStateMachineServiceClient.cs
@@ -1,0 +1,7 @@
+﻿namespace REstate.Remote.Services
+{
+    public interface IStateMachineServiceClient
+    {
+        IStateMachineService Create();
+    }
+}
