+++ src/REstate.Remote/Services/IStateMachineServiceLocalAdapter.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using REstate.Remote.Models;
+using REstate.Schematics;
+
+namespace REstate.Remote.Services
+{
+    internal interface IStateMachineServiceLocalAdapter
+    {
+        Task BulkCreateMachineFromSchematicAsync<TState, TInput>(Schematic<TState, TInput> schematic, IEnumerable<IDictionary<string, string>> metadata, CancellationToken cancellationToken = default);
+        Task BulkCreateMachineFromStoreAsync<TState, TInput>(string schematicName, IEnumerable<IDictionary<string, string>> metadata, CancellationToken cancellationToken = default);
+        Task<CreateMachineResponse> CreateMachineFromSchematicAsync<TState, TInput>(Schematic<TState, TInput> schematic, string machineId, IDictionary<string, string> metadata, CancellationToken cancellationToken = default);
+        Task<CreateMachineResponse> CreateMachineFromStoreAsync<TState, TInput>(string schematicName, string machineId, IDictionary<string, string> metadata, CancellationToken cancellationToken = default);
+        Task DeleteMachineAsync<TState, TInput>(string machineId, CancellationToken cancellationToken);
+        Task<GetMachineMetadataResponse> GetMachineMetadataAsync<TState, TInput>(string machineId, CancellationToken cancellationToken = default);
+        Task<GetMachineSchematicResponse> GetMachineSchematicAsync<TState, TInput>(string machineId, CancellationToken cancellationToken = default);
+        Task<GetSchematicResponse> GetSchematicAsync<TState, TInput>(string schematicName, CancellationToken cancellationToken);
+        Task<SendResponse> SendAsync<TState, TInput>(string machineId, TInput input, Guid? commitTag, CancellationToken cancellationToken = default);
+        Task<SendResponse> SendWithPayloadAsync<TState, TInput, TPayload>(string machineId, TInput input, TPayload payload, Guid? commitTag, CancellationToken cancellationToken = default);
+        Task<StoreSchematicResponse> StoreSchematicAsync<TState, TInput>(Schematic<TState, TInput> schematic, CancellationToken cancellationToken = default);
+    }
+}
